requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
pymongo>=4.5.0
motor>=3.3.1
python-jose>=3.3.0
passlib>=1.7.4
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
python-multipart>=0.0.6
bcrypt>=4.0.1
email-validator>=2.0.0
starlette>=0.27.0
httpx>=0.24.1 